refactor: (#31) constants, types, utils index.ts 참조하도록 수정
