+++ src/hooks/usePassword.ts
@@ -1,5 +1,5 @@
 import { useInputState } from '@/hooks';
-import { passwordRegex } from '@/constants/regex.constants';
+import { passwordRegex } from '@/constants';
 import { useEffect } from 'react';
 
 interface UsePasswordReturn {
