+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export * from './api';
+export * from './auth';
+export * from './comment';
+export * from './webtoon';
+export * from './webtoonDetail';
+export * from './webtoonViewer';
